// src/objects/common/enums/Feeder.ts
enum Feeder {
  TETUAN = "Tetuan",
  MERCEDES = "Mercedes",
  NORTHSIDE = "Northside",
  ZAMPEN = "Zampen",
  GUIWAN = "Guiwan",
  TUMAGA = "Tumaga",
  ARENA_BLANCO = "Arena Blanco",
  PRESA = "Presa",
  SAN_JOSE = "San Jose",
  CALARIAN = "Calarian",
  BALIWASAN = "Baliwasan",
  DEDICATED_LINE = "Dedicated Line",
  INTERCO = "Interco",
  CULIANAN = "Culianan",
  CURUAN = "Curuan",
  FISHING_PORT = "Fishing Port",
  TALISAYAN = "Talisayan",
  LABUAN = "Labuan",
  RECODO = "Recodo",
  SINUNUC = "Sinunuc",
  BIG_FISH = "Big Fish",
  TIN_CAN_ORIENTAL = "Tin Can Oriental (1-2MVA)",
  MEGA_FISHIN_CORP = "Mega Fishing Corp (1-4MVA)",
  MEGA_FISHIN_CORP2 = "Mega Fishing Corp (1-3MVA)",
  PASONANCA = "Mega Fishing Corp (1-3MVA)",
  SAN_ROQUE = "San Roque",
  ACLEM = "Aclem",
  GOVERNOR_ALVAREZ = "Governor Alvarez",
  VETERANS_AVENUE = "Veterans Avenue",
  NUNEZ = "Nunez",
  STA_CATALINA = "Sta Catalina",
  RIO_HONDO = "Rio Hondo",
  GUARDIA_NACIONAL = "Guardia Nacional",
  CAMPANER = "Campaner",
  KCC_MALL = "KCC Mall (1-10MVA)",
}

export default Feeder;